Move MoonPhase and Enumeration to main project
