+++ PgMoon-Plugin/Data/Enumeration.cs
@@ -1,0 +1,42 @@
+namespace PgMoon.Data
+{
+    using System;
+    
+    
+    public abstract class Enumeration : IComparable
+    {
+        public string Name { get; private set; }
+
+        public int Id { get; private set; }
+
+        protected Enumeration(int id, string name) => (Id, Name) = (id, name);
+
+        public override string ToString() => Name;
+
+        // override object.Equals
+        public override bool Equals(object obj)
+        {
+            //
+            // See the full list of guidelines at
+            //   http://go.microsoft.com/fwlink/?LinkID=85237
+            // and also the guidance for operator== at
+            //   http://go.microsoft.com/fwlink/?LinkId=85238
+            //
+
+            if (obj == null || GetType() != obj.GetType())
+            {
+                return false;
+            }
+
+            Enumeration that = (Enumeration)obj;
+            return this.Id == that.Id;
+        }
+
+        public override int GetHashCode() => Id;
+
+        public int CompareTo(object other) => Id.CompareTo(((Enumeration)other).Id);
+
+        // Other utility methods ...
+    }
+   
+}
